% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpcr.R
\name{gwpcr}
\alias{dgwpcr}
\alias{dgwpcr.fun}
\alias{gwpcr}
\alias{pgwpcr}
\alias{pgwpcr.fun}
\alias{rgwpcr}
\title{PCR Product Distribution induced by a Binomial Galton-Watson Process}
\usage{
rgwpcr(n, efficiency, molecules = 1, cycles = NA)

dgwpcr(l, efficiency, molecules = 1)

dgwpcr.fun(efficiency, molecules = 1)

pgwpcr.fun(efficiency, molecules = 1)

pgwpcr(l, efficiency, molecules = 1)
}
\arguments{
\item{n}{number of random samples to generate}

\item{efficiency}{efficiency of amplification}

\item{molecules}{initial copy number}

\item{cycles}{number of amplification cycles used for simulation. By default,
a large enough value is used to make the results virtually
idistinguishable from the limit for \eqn{cycles \to \infty}}

\item{l}{molecular family size relative to the average}
}
\description{
Write Me
}

