% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpcr.R
\name{gwpcr.sd}
\alias{gwpcr.sd}
\alias{gwpcr.sd.inv}
\title{PCR Product Distribution Standard Deviation}
\usage{
gwpcr.sd(efficiency, molecules = 1)

gwpcr.sd.inv(sd, molecules = 1)
}
\arguments{
\item{efficiency}{efficiency of amplification}

\item{molecules}{initial copy number}

\item{sd}{Standard deviation of the PCR product distribution}
}
\description{
For efficiency \var{E} and initial number of molecules \var{m},
  the PCR product distribution has \emph{variance} (not standard deviation!)
  \eqn{\frac{1-E}{1+E}\cdot\frac{1}{m}}{(1+E) / ((1-E) * m)}.

  Function \code{gwpcr.sd} uses this to compute the standard deviation (i.e.
  the square of the above) for a given efficiency and initial copy number,
  and \code{gwpcr.sd.inv} does the reverse and computes the efficiency given
  standard deviation and copy number.
}
\seealso{
\code{\link{gwpcr}}
}

