% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{gwpcrpois.est}
\alias{gwpcrpois.est}
\title{Parameter Estimation for PCR-Poisson Mixture}
\usage{
gwpcrpois.est(x = NULL, mean = NULL, var = NULL, n.umis = NULL,
  method = "mom", must.converge = TRUE, threshold = 1, molecules = 1,
  loss = expression(p0), ctrl = list())
}
\arguments{
\item{x}{a numeric vector containing the observed read counts per molecular
family, after removal of families below the detection threshold. The
vector must thus contain only whole numbers not smaller than \var{threshold}.
This parameter and the combination of parameters \var{mean} and \var{var}
are mutually exclusive.}

\item{mean}{average number of observations per molecular family
\emph{computed over the unambiguously detected famililies}, i.e. over those
families which were observed at least \var{threshold} times. This parameter
and specifying an observation vector through parameter \var{x} are mutually
exclusive, and specifying \var{mean} and \var{var} instead of the full
observation vector is only possible for estimation method 'mom'.}

\item{var}{standard deviations of number of observations per molecular
family, also \emph{computed over the unambiguously detected famililies},
i.e. over those families which were observed at least \var{threshold} times.
This parameter and specifying an observation vector through parameter \var{x}
are mutually exclusive, and specifying \var{mean} and \var{var} instead of
the full observation vector is only possible for estimation method 'mom'.}

\item{n.umis}{the number of observed \acronym{UMI}s, used in the estimation
of \code{n.tot} (i.e. the total number of molecules/\acronym{UMI}s in the
original sample). See also the dicussion of the parameter \code{loss}, and
the result value \code{n.tot}.}

\item{method}{the estimation method to use, either 'mle' for \emph{maximum
likelihood estimation} or 'mom' for \emph{method of moments}. See Details.}

\item{must.converge}{if set to \code{TRUE}, an error is reported of the
parameter estimation fails to converge. If \code{FALSE}, a warning is
reported instead.}

\item{threshold}{minimal number of observations a molecular family must have
to count as \dfn{unambiguously detected}. Setting this to a value \eqn{v >=
0} \emph{conditions} the distribution on \eqn{c >= c}, i.e every value of
\eqn{c} less than that gets assigned probability zero.}

\item{molecules}{initial copy number}

\item{loss}{an expression specifying how the loss, i.e. the percentage of
all molecules (or UMIs) that was not observed, or removed by the read
count threshold. In the simple case of each read count observation
representing a separate molecules, the default value \var{p0} is correct
-- the lost molecules are then simply those whose read count lies below
the specified \var{threshold}. In more complex scenarios, e.g. if a
single molecule produces separate read count for each strand, which are
then either both rejected or both accepted, the additional rejection cases
must be considered by a custom loss expression}

\item{ctrl}{a list of settings controlling the estimation procedure.
Difference estimation methods recognize different possible \var{ctrl}
settings, unrecognized settings are ignored without warning. See Details
for the settings relevant to each estimation method.}
}
\value{
A list containing the values

  \item{convergence}{flag indicating whether the estimation converged.
  \var{0} indicates convergence.}

  \item{efficiency}{parameter estimate for \var{efficiency} (see
  \link{gwpcrpois})}

  \item{lambda0}{parameter estimate for \var{lambda0} (see \link{gwpcrpois})}

  \item{p0}{probability of observing a read count less than the specified
  \var{threshold}}

  \item{loss}{the estimation loss according to the specified loss expression,
   i.e. percentage of molecules not observed or filtered out}

  \item{n.tot}{the estimated total number of molecules in the sample, i.e.
  \code{n.ums / (1 - loss)}}

  \item{n.obs}{The length of the observation vector \var{x} used for parameter
  estimation, or \code{NA} if \var{mean} and \var{var} were specified
  directly.}

  \item{n.umis}{The number of observed molecules/\acronym{UMI}s specified in
  the call to \code{gwpcrpois.est}}

  \item{threshold}{detection threshold specified in the call to
  \code{gwpcrpois.est}}

  \item{molecules}{initial molecule count specified in the call to
  \code{gwpcrpois.est}}
}
\description{
Estimates the parameters \var{efficiency} and \var{lambda0} from a vector
of observed read counts per molecular family, or (depending on the estimation
method) the mean and variance of these observations. Supports arbitrary
detection thresholds and initial molecule counts, but estimation may be
considerably faster in the (unrealistic) case \var{threshold=0} than in the
general one.
}
\details{
The two available estimation methods, \emph{method of
  moments} (\code{method='mome'}) and \emph{maximum likelihood}
  (\code{method='mle'}) have different propertiers and accept different
  \var{ctrl} parameters:
  \describe{
  \item{method of moments (\code{mom}):}{
  For the (unrealistic) uncensored case, i.e. \var{threshold=0}, the
  specified mean is the method-of-moments estimate for \var{lambda0}, and a
  closed formula is used to compute \var{efficiency} from \var{mean} and
  \var{var}. The \var{ctrl} argument is not used in this case.

  In case of a censored distribution, the \sample mean is not a consistent
  estimator for \var{lambda0} because the expectation of the censored
  distribution is in general larger than \var{lambda0}. The sample variance
  simiarly deviates from the variance of the uncensored distribution.

  An interative approach is used to find method-of-moment estimates in this
  case. Initial estimates are computed as if \var{threshold} were zero. From
  these the probability \var{pdetect} (i.e. \code{1-p0}) of detecting a particular family is
  found, and used to correct for the biases in the sample mean and variance.
  Then the parameter estimates are updated. This process continues until it
  either converges or reaches the maximum allowed number of iterations. Both
  termination criteria can be controlled via the \var{ctrl} parameter, which
  is a list that can contain the following components: \describe{

  \item{maxit}{Maximum number of iterations. Defaults to 150.}

  \item{rel.tol}{Relative convergence tolerance. Applied to \var{efficiency},
  \var{lambda0} and the detection probability \var{pdetect}. Defaults to
  \code{1e-4}.}

  \item{rel.tol}{Absolute convergence tolerance. Only used as the tolerance
  around zero, where the relative tolerance becomes meaningless. Defaults to
  \code{1e-4}.}

  \item{trace}{Output estimates after each round}
  }
  }
  \item{maximum likelihood (\code{mle}):}{
  The parameters are estimated by maximizing the log-likelihood using
  \code{\link{optim}}. The \var{ctrl} settings are passed through to
  \code{\link{optim}}, except for \var{fnscale} and \var{parscale} which
  are overwritten. The method of moments (method \code{'mom'})) estimate
  is used as the starting point during liklihood optimization.
  }
  }
}
\seealso{
\code{\link{gwpcrpois}}
}
