% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{gwpcrpois.groupest}
\alias{gwpcrpois.groupest}
\title{Group-wise Parameter Estimation for PCR-Poisson Mixture}
\usage{
gwpcrpois.groupest(formula, data, method = "mom", threshold = 1,
  molecules = 1, loss = expression(p0), ctrl = list())
}
\arguments{
\item{formula}{Formula of the form \code{reads ~ key1 + key2 + ...},
where \var{reads} is the column containing the per-UMI read count,
and \var{key1}, \var{key2}, ... are the column(s) that uniquely identify
a group. If multiple read counts are observed per UMI (for example if a
protocol that yields strand-specific counts for the two strands of double-
stranded molecules is used), \code{c(reads1, reads2, ...)} can be used as
the left hand side to combine read counts from multiple columns. Note that
in this case, the default \var{loss} expression is probably not appropriate.
All the read counts in the columns listed in \var{reads} must be greater or
equal than \var{threshold}.}

\item{data}{a \code{\link{data.frame}} or \code{\link{data.table}} with one
row per observed UMI. The required columns are determined by the \var{formula}.}

\item{method}{the estimation method to use, either 'mle' for \emph{maximum
likelihood estimation} or 'mom' for \emph{method of moments}. See Details.}

\item{threshold}{minimal number of observations a molecular family must have
to count as \dfn{unambiguously detected}. Setting this to a value \eqn{v >=
0} \emph{conditions} the distribution on \eqn{c >= c}, i.e every value of
\eqn{c} less than that gets assigned probability zero.}

\item{molecules}{initial copy number}

\item{loss}{an expression specifying how the loss, i.e. the percentage of
all molecules (or UMIs) that was not observed, or removed by the read
count threshold. In the simple case of each read count observation
representing a separate molecules, the default value \var{p0} is correct
-- the lost molecules are then simply those whose read count lies below
the specified \var{threshold}. In more complex scenarios, e.g. if a
single molecule produces separate read count for each strand, which are
then either both rejected or both accepted, the additional rejection cases
must be considered by a custom loss expression}

\item{ctrl}{a list of settings controlling the estimation procedure.
Difference estimation methods recognize different possible \var{ctrl}
settings, unrecognized settings are ignored without warning. See
\code{\link{gwpcrpois.est}} for the \var{ctrl} settings affecting the raw
group-specific estimates, and the Details section for settings affecting
the shrinkage of those estimates.}
}
\value{
a \code{\linkS4class{data.table}} containing one row per group, and
  which besides the group key column(s) as indicated by the \var{formula},
  contains the following columns

  \item{efficiency}{the final group-specific estimate of the PCR efficiency}

  \item{lambda0}{the final group-specific estimate of the number of reads per
  molecule}

  \item{loss}{the final group-specific estimate of the fraction of lost
  molecules.}

  \item{n.tot}{the final group-specific estimate of the total (loss-corrected)
  number of molecules.}
}
\description{
Estimates the parameters \var{efficiency} and \var{lambda0} separately for each
\emph{group} of UMIs (i.e, for example separately for each gene, each exon, each
cell, ...). To reduce the noise of the raw group-specific parameters, the raw
group-specific parameters are \emph{shrunken} towards global estimates for
\var{efficiency} and \var{lambda0} that are specified
}
\details{
Apart from the \var{ctrl} parameters described in
  \code{\link{gwpcrpois.est}}, the following settings can be modified
  \describe{
  \item{core}{number of CPU cores to use to compute group-specific estimates.
  If set to a value greather than 1, the \code{\link[=pkg]{parallel}} package
  must be loaded.}

  \item{obs.min.ingroup}{how many observed per-UMI read counts a group must
  contain for group-specific estimates to be computed and used. Default is 5.}

  \item{use.nonconv.globalest}{whether to use global estimates that didn't
  fully converge (i.e. where \code{\link{gwpcrpois.est}} reported a value other
  than \code{0} for \var{convergence}). Default is \code{FALSE}.}

  \item{use.nonconv.groupest}{whether to use group-specific estimates
  that didn't fully converge (i.e. where \code{\link{gwpcrpois.est}} reported
  a value other than \code{0} for \var{convergence}). Default is \code{FALSE}.}

  \item{include.mean.var}{whether to include columns with the group-wise
  global mean and variance of readcounts in the output table. Default is
  \code{FALSE}.}

  \item{verbose}{whether to output progress and diagnostic messages, default
  is \code{FALSE}.}
  }
}
\seealso{
\code{\link{gwpcrpois}}
}
