% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{gwpcrpois.mle}
\alias{gwpcrpois.mle}
\title{Compatibility wrapper of \code{\link{gwpcrpois.est}}}
\usage{
gwpcrpois.mle(c, threshold = 1, molecules = 1)
}
\description{
Compatibility wrapper of \code{\link{gwpcrpois.est}}
}
\seealso{
\code{\link{gwpcrpois.est}}
}
