% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{gwpcrpois.mom.groupwise}
\alias{gwpcrpois.mom.groupwise}
\title{Compatibility wrapper of \code{\link{gwpcrpois.est}}}
\usage{
gwpcrpois.mom.groupwise(formula, data, threshold = 1, molecules = 1,
  loss = expression(p0), ctrl = list())
}
\description{
Compatibility wrapper of \code{\link{gwpcrpois.est}}
}
\seealso{
\code{\link{gwpcrpois.est}}
}
