% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{gwpcrpois.mom}
\alias{gwpcrpois.mom}
\title{Compatibility wrapper of \code{\link{gwpcrpois.est}}}
\usage{
gwpcrpois.mom(mean, var, threshold = 1, molecules = 1, ctrl = list(),
  nonconvergence.is.error = FALSE)
}
\description{
Compatibility wrapper of \code{\link{gwpcrpois.est}}
}
\seealso{
\code{\link{gwpcrpois.est}}
}
