% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpcrpois.R
\name{gwpcrpois}
\alias{dgwpcrpois}
\alias{gwpcrpois}
\alias{pgwpcrpois}
\alias{rgwpcrpois}
\title{Poissonian Sampling Distribution of the PCR Product Distribution}
\usage{
rgwpcrpois(samples, efficiency, lambda0, threshold = 1, molecules = 1)

dgwpcrpois(c, efficiency, lambda0, threshold = 1, molecules = 1,
  clamp.efficiency = TRUE)

pgwpcrpois(c, efficiency, lambda0, threshold = 1, molecules = 1,
  clamp.efficiency = TRUE)
}
\arguments{
\item{efficiency}{}

\item{molecules}{}

\item{n}{}

\item{lambda}{}
}
\description{
Test
}
\seealso{
gwpcr
}

