% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpcr.R
\name{gwpcr}
\alias{dgwpcr}
\alias{dgwpcr.fun}
\alias{gwpcr}
\alias{pgwpcr}
\alias{pgwpcr.fun}
\alias{rgwpcr}
\title{Binomial Galton-Watson Model for the Polymerase Chain Reaction (PCR)}
\usage{
rgwpcr(n, efficiency, molecules = 1, cycles = NA)

dgwpcr(l, efficiency, molecules = 1)

dgwpcr.fun(efficiency, molecules = 1)

pgwpcr.fun(efficiency, molecules = 1)

pgwpcr(l, efficiency, molecules = 1)
}
\arguments{
\item{n}{number of random samples to generate}

\item{efficiency}{efficiency of amplification}

\item{molecules}{initial copy number}

\item{cycles}{number of amplification cycles used for simulation. By default,
a large enough value is used to make the results virtually idistinguishable
from the limit for \eqn{cycles \to \infty}}

\item{l}{molecular family size relative to the average}
}
\description{
Limit distribution of amount of PCR product relative to the
  expected value after many cycles as prediced by a binomial Galton-Watson
  model of the polymerase chain reaction (PCR), Parameters are
  \var{efficiency} and \var{molecules} -- the former is the probability that
  a molecule is duplicated during a particular PCR cycle, the latter the
  initial number of molecules in the reaction mix.

  Note that a \emph{molecule} refers to single strand here, and complementary
  strands are not distinguished. Setting \var{molecules=2} thus models a PCR
  reaction starting from a single piece of double-stranded DNA.

  \code{dgwpcr} (resp. \code{pgwpcr}) evaluate the density (resp. the CDF) at
  the given relative amount of PCR product \code{l} for parameters
  \var{efficiency} and \var{molecules}. \code{dgwpcr.fun} (resp.
  \code{pgwpcr.fun}) return a unary function which represents the density
  (resp. CDF) for the given parameters. \code{rgwpcr} draws random samples by
  simulation. If the number of PCR cycles to use is not specified, the
  simulation is stopped once the expected absolute amount of PCR product
  reaches one million molecules, at which point the distribution is
  considered to be close to the limit distribution for infinitly many cycles.
}
\details{
The binomial Galton-Watson PCR model treats PCR as a branching
  process. At time 0, the absolute number of molecules \eqn{c_n} is the
  initial copy number \var{molecules}. Each time step from \eqn{c_n} to
  \eqn{c_{n+1}}{c_(n+1)} corresponds to a PCR cycle and duplicates each of
  the \eqn{c_n} molecules with the probability specified in parameter
  \var{efficiency} (\var{E}). Thus,

  \deqn{c_{n+1} = c_n + \textrm{Binomial}(c_n, E).}{c_(c+1) = c_n +
  Binomial(c_n, E).}

  Each cycle thus increases the expected molecule count by a factor of
  \eqn{(1+E)}. The \emph{relative} amount of PCR product after \eqn{n} cycles
  is therefore

  \deqn{l_n = c_n \cdot (1+E)^{-n}.}{l_n = c_n (1+E)^-n.}

  \code{dgwpcr} (resp. \code{rgwpcr}) is the density (resp. CDF) of the a.c.
  limit in distribution of \eqn{l_n} for \eqn{n \to \infty}{n to \infty}.
  Essentially, the reason that \eqn{l_n} converges in distribution is that
  the larger the number of molecules \eqn{c_n}, the smaller the additional
  variability introduced into \eqn{c_{n+1}}{c_(n+1)} by the term
  \eqn{\textrm{Binomial}(c_n, E).}{Binomial(c_n, E).}. For reasonably large
  efficiencies, that becomes quickly negligible compared to \eqn{c_n}. Assume
  e.g. an efficiency of 50\%, and that \eqn{c_n=10^4}{c_n=2500}. The standard
  deviation of the binomial term is then already only 25, i.e. 1\% of
  \eqn{c_n}, and this state can be expected to be reached after about 20
  cycles.

  For this reason, using the limit distribution still yields a model of the
  polymerase chain reaction that is sufficiently accurate for most purposes.
}
\seealso{
\code{\link{gwpcr.sd}}

\code{\link{gwpcr.mixture}}

\code{\link{gwpcrpois}}
}

