% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwpcr.R
\name{gwpcr}
\alias{dgwpcr}
\alias{dgwpcr.fun}
\alias{gwpcr}
\alias{pgwpcr}
\alias{pgwpcr.fun}
\alias{rgwpcr}
\title{PCR Product Distribution induced by a Binomial Galton-Watson Process}
\usage{
rgwpcr(n, efficiency, molecules = 1, cycles = NA)

dgwpcr(lambda, efficiency, molecules = 1)

dgwpcr.fun(efficiency, molecules = 1)

pgwpcr.fun(efficiency, molecules = 1)

pgwpcr(lambda, efficiency, molecules = 1)
}
\arguments{
\item{n}{}

\item{efficiency}{}

\item{molecules}{}

\item{lambda}{}
}
\description{
Test
}

