% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{gwpcrpois.mle}
\alias{gwpcrpois.mle}
\title{Maximum-Likelihood Parameter Estimation for PCR-Poisson Mixture}
\usage{
gwpcrpois.mle(c, threshold = 1, molecules = 1)
}
\arguments{
\item{c}{number of observations of a particular molecular family}

\item{threshold}{minimal number of observations a molecular family must
have to count as observed. Setting this to a value \eqn{v >= 0}
\emph{conditions} the distribution on \eqn{c >= c}, i.e
every value of \eqn{c} less than that gets assigned probability
zero.}

\item{molecules}{initial copy number}
}
\description{
Maximum-Likelihood Parameter Estimation for PCR-Poisson Mixture
}

