% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{gwpcrpois.mom.groupwise}
\alias{gwpcrpois.mom.groupwise}
\title{Shrinkage Estimator for Individual Parameters of PCR-Poisson Mixture}
\usage{
gwpcrpois.mom.groupwise(formula, data, threshold = 1, molecules = 1,
  ctrl = list())
}
\arguments{
\item{threshold}{minimal number of observations a molecular family must have
to count as \dfn{unambiguously detected}. Setting this to a value \eqn{v >=
0} \emph{conditions} the distribution on \eqn{c >= c}, i.e every value of
\eqn{c} less than that gets assigned probability zero.}

\item{molecules}{initial copy number}
}
\description{
Low-noise estimates of \var{efficiency} and \var{lambda0} for subsets of the
data (e.g.' for individual genes of an RNA-Seq experiment), even if those
subsets contain only few UMIs.
}
\details{
XXX
}
