% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{gwpcrpois.mom}
\alias{gwpcrpois.mom}
\title{Method-of-Moments Parameter Estimation for PCR-Poisson Mixture}
\usage{
gwpcrpois.mom(mean, var, threshold = 1, molecules = 1)
}
\arguments{
\item{mean}{average number of observations per molecular family
\strong{computed over the actually observed famililies}, i.e.
over those families which were observed at least \code{threshold}
times.}

\item{var}{standard deviations of number of observations per molecular
family, also \strong{computed over the actually observed famililies}, i.e.
 over those families which were observed at least \code{threshold}
 times.}

\item{threshold}{minimal number of observations a molecular family must
have to count as observed. Setting this to a value \eqn{v >= 0}
\emph{conditions} the distribution on \eqn{c >= c}, i.e
every value of \eqn{c} less than that gets assigned probability
zero.}

\item{molecules}{initial copy number}
}
\description{
Estimates the parameters \code{efficiency} and \code{lambda0} from the
observed mean and variance. Supports estimation using a censored (conditional)
distribution, i.e. the case where a molecular family is only considered
to be observed if it is observed at least \code{threshold} number of times.
}

