% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{gwpcrpois.mom}
\alias{gwpcrpois.mom}
\title{Method-of-Moments Parameter Estimation for PCR-Poisson Mixture}
\usage{
gwpcrpois.mom(mean, var, threshold = 1, molecules = 1, ctrl = list())
}
\arguments{
\item{mean}{average number of observations per molecular family
\emph{computed over the unambiguously detected famililies}, i.e. over those
families which were observed at least \var{threshold} times.}

\item{var}{standard deviations of number of observations per molecular
family, also \emph{computed over the unambiguously detected famililies},
i.e. over those families which were observed at least \var{threshold}
times.}

\item{threshold}{minimal number of observations a molecular family must have
to count as \dfn{unambiguously detected}. Setting this to a value \eqn{v >=
0} \emph{conditions} the distribution on \eqn{c >= c}, i.e every value of
\eqn{c} less than that gets assigned probability zero.}

\item{molecules}{initial copy number}
}
\value{
A list containing the values

  \item{convergence}{flag indicating whether the estimation converged.
  \code{0} indicates convergence.}

  \item{efficiency}{parameter estimate for \var{efficiency} (see
  \link{gwpcrpois})}

  \item{lambda0}{parameter estimate for \var{lambda0} (see \link{gwpcrpois})}

  \item{pdetect}{probability of unambiguosly detecting a particular molecular
  family, i.e of it having at least \var{threshold} observations}
  
  \item{p0}{probability of not unambiguosly detecting a particular molecular
  family, i.e of it having fewer than \var{threshold} observations. Always
  equal to \code{1-pdetect}}

  \item{threshold}{detection threshold specified in the call to
  \code{gwpcrpois.mom}}

  \item{molecules}{initial molecule count specified in the call to
  \code{gwpcrpois.mom}}
}
\description{
Estimates the parameters \var{efficiency} and \var{lambda0} from the sample
mean and variance. Supports arbitrary detection thresholds and initial
molecule counts, but estimation is considerably faster in the (unrealistic)
case \var{threshold=0} than in the general one.
}
\details{
For the (unrealistic) uncensored case, i.e. \var{threshold=0}, the
  specified mean is the method-of-moments estimate for \var{lambda0}, and a
  closed formula is used to compute \var{efficiency} from \var{mean} and
  \var{var}. The \var{ctrl} argument is not used in this case.

  In case of a censored distribution, the sample mean is not a consistent
  estimator for \var{lambda0} because the expectation of the censored
  distribution is in general larger than \var{lambda0}. The sample variance
  simiarly deviates from the variance of the uncensored distribution.

  An interative approach is used to find method-of-moment estimates in this
  case. Initial estimates are computed as if \var{threshold} were zero. From
  these the probability \var{pdetect} of detecting a particular family is
  found, and used to correct for the biases in the sample mean and variance.
  Then the parameter estimates are updated. This process continues until it
  either converges or reaches the maximum allowed number of iterations. Both
  termination criteria can be controlled via the \var{ctrl} parameter, which
  is a list that can contain the following components: \describe{

  \item{maxit}{Maximum number of iterations. Defaults to 150.}

  \item{rel.tol}{Relative convergence tolerance. Applied to \var{efficiency},
  \var{lambda0} and the detection probability \var{pdetect}. Defaults to
  \code{1e-4}.}

  \item{rel.tol}{Absolute convergence tolerance. Only used as the tolerance
  around zero, where the relative tolerance becomes meaningless. Defaults to
  \code{1e-4}.}

  \item{trace}{Output estimates after each round}

  }
}
\seealso{
\code{\link{gwpcrpois}}
}
